refactored config.service to set up for all 3 customizers
