+++ apps/kamui-customizer/src/app/app.module.ts
@@ -3,16 +3,18 @@ import { NgModule } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
 import { FeCustomizerEngineModule } from 'fe-customizer-engine';
 import { AppComponent } from './app.component';
+import { KamuiComponent } from './routes/kamui/kamui.component';
+import { KamuiCustomizerComponent } from './component/kamui-customizer/kamui-customizer.component';
 
 @NgModule({
-  declarations: [AppComponent],
+  declarations: [AppComponent, KamuiComponent, KamuiCustomizerComponent],
   imports: [BrowserModule, FeCustomizerEngineModule],
   providers: [
     {
       provide: APP_BASE_HREF,
       useFactory: (s: PlatformLocation) => s.getBaseHrefFromDOM(),
-      deps: [PlatformLocation]
-    }
+      deps: [PlatformLocation],
+    },
   ],
   bootstrap: [AppComponent],
 })
