+++ apps/kamui-customizer/src/app/component/kamui-customizer/kamui-customizer.component.html
@@ -1,0 +1,1 @@
+<p>kamui-customizer works!</p>
