+++ apps/kamui-customizer/src/app/component/kamui-customizer/kamui-customizer.component.ts
@@ -1,0 +1,12 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'fecustomizer-engine-kamui-customizer',
+  templateUrl: './kamui-customizer.component.html',
+  styleUrls: ['./kamui-customizer.component.scss'],
+})
+export class KamuiCustomizerComponent implements OnInit {
+  constructor() {}
+
+  ngOnInit(): void {}
+}
