+++ apps/kamui-customizer/src/app/config/config-keys.ts
@@ -1,0 +1,5 @@
+export enum ConfigKey{
+    Corrin = 'CORN',
+    Robin = 'ROBIN',
+    Kris = 'KRIS'
+}
