+++ apps/kamui-customizer/src/app/config/index.ts
@@ -1,0 +1,2 @@
+export * as corrinConfigs from './corn.config';
+export * from './config-keys';
