+++ apps/kamui-customizer/src/app/routes/kamui/kamui.component.html
@@ -1,0 +1,1 @@
+<p>kamui works!</p>
