+++ apps/kamui-customizer/src/app/routes/kamui/kamui.component.ts
@@ -1,0 +1,12 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'fecustomizer-engine-kamui',
+  templateUrl: './kamui.component.html',
+  styleUrls: ['./kamui.component.scss'],
+})
+export class KamuiComponent implements OnInit {
+  constructor() {}
+
+  ngOnInit(): void {}
+}
